move RmlSolLua out of the lib folder so that it can integrate more with the rest of the Spring engine
