+++ rts/Rml/SolLua/RmlSolLua.cpp
@@ -1,4 +1,6 @@
-﻿#include "RmlSolLua/RmlSolLua.h"
+﻿// This file is part of the Spring engine (GPL v2 or later), see LICENSE.html
+
+#include "RmlSolLua.h"
 
 #include <sol2/sol.hpp>
 #include <RmlUi/Core.h>
