+++ rts/Rml/SolLua/TranslationTable.h
