+++ rts/Rml/SolLua/bind/Colour.cpp
