+++ rts/Rml/SolLua/bind/Log.cpp
