--- rts/lib/RmlSolLua/.gitignore
@@ -1,3 +1,0 @@
-.vs
-out
-!.gitignore
